PyCraft 3.0 Demo
WARNING: This is a demo version of PyCraft 3.0 and is unfinished, buggy, and missing lots of features.

The demo for PyCraft 3.0 is finally here! PyCraft 3.0 Demo introduces chunking, allowing for infinite worlds and increased performance. Chunking currently does not save changes or destruction in terrain or placed blocks. This will be added in the release version of PyCraft 3.0.